Bring the load performance in line with what is was previously.
Before this change, the branch took 42s vs 28s on master to load.

The problem was that wordCountsForSemantics was called twice instaed of once. The reason was that sometimes it was called with an _explicit_ undefined at the end of the arguments, and sometimes with a missing undefiend.

Changing arrayEqual to ignore trailing undefined when comparing for equality fixes this cache miss.

This now makes the branch also around 28s.

Part of #694.