Add an optimnization to fontSizeBoosts to skip calculating if that field hasn't changed.
This gives a significant (~30%) speedup when multi-editing hundreds of cards. Before each fontSizeBoost was calculated serially for each modified card, even though its content that could affect font sizing hadn't changed.

fontSizeBoosts is pretty expensive, but if it's done for a single card it's not that big of a deal. It's when it's done sequentially for a ton of cards at once that it's a problem.

Part of #701.