Actually wire in the modification enqueing.
The idea that if we're expecting modifications, instead of applying them immediately we enque them. Then once all of the expected ones have been received, we apply them all at once.

Note a few gaps in how this works: 1) if card updates come in that are NOT the ones you were expecting it will apply them early, and 2) if updates come in from multiple different CardFetchTypes there might be up to n (where n is number of CardFetchTypes) of serial applications. But this is OK because it's just an optimization and that should be rare.

Before this, if you edited more cards than fit in one Batch, you'd get multiple sequential expensive updates of cards as each batch landed, which could make it take a very long time.

Part of #701.