A first rough stab at counting words for spell checking.
I haven't checked how much slower this makes loading / editing or anything, this was more to get a sense for how effective this technique might be.

Unfortunately the answer seems to be "not very" in spot checking these lists. For example, "clover" shows up a single time in the corpus as does "faculty". That implies the false positive rate for misspellings will be huge.

part of #651.