Add "action" creators for adding and removing dictionary overriddes.
Note that they will be run as soon as a user does them while editing, NOT waiting to do it until the card is saved.

This is more akin to how we handle creating new tags or slugs; they happen live while editing, and even if you x out of editing they still happened.

This is both simpler to implement and also a reasonable behavior.

Part of #651.