Switch the storage scheme for ids in dictionary overrides.
Before they had ids of a short bit of random text. But that made it eyeballing the list annoying and also meant that the actual document had a non-deterministic id that had to be kept track of... and also made it easier to inadvertnetly save a duplictae.

Now the id is the normalizedFirestoreID of the word, and becuase every so often that won't be precisely the word, the actual full normalized word is stored in word.

Part of #651.