Create a Timestamp interface that can be used in both server and client.
Importing a real Timestamp in funtiions and client is different (firebase vs firebase-admin), which makes sharing a lot of types that use it hard.

This craetes an interafce that has the expected methods that can be used instead of a real timestamp to allow more types to be moved to shared.

Not actually used anwyhere yet.

Part of #134.