Simplify it so chat_message_chunks are a top level collection.
Simply store the owner ID in the chunk, since they are the only ones that should be able to see it.

Having it in a top-level function will make periodic cleaning of it easier, since chunks should be short lived, and could be a bunch of duplicated storage.

Part of #710.