Refactor it so the assistant message is not kicked off by createChat or postMessage.
Instead, it is initated by the client when it receives an assistant message in status 'ready'.

This means it's up to a client to notice there's a new message thta needs to be streamed and tickle the server to stream it. As soon as it's tickled, it's changed to 'streaming' as a kind of poor mans lock to make it less likely two dueling streaming are kicked off.

Right now, it's just a tickle that gets the completion and writes it to firestore, but this will make it much easier to add actual streaming of tokens to the client. Note that this architecture prevents multiple clients from seeing the in-progress streaming tokens, which is fine.

This design also makes it easier fo ra client to show a 'retry' UI and make it easier to ahve a signaling architecture to retry failed messages.

We split out streamMessages because otherwise we'd need createChat and postMessage to return messages in an event stream, which is awkward. Now the only endpoint that will need to retrun messages in an event stream is the one that actually is a stream.

The next step is to actually have streamMessage stream the in progress tokens.

Part of #710.