Switch permissions fetching logic to use a snapshot, not getDoc.
The logic was finicky when firestore thinks it's disconnected (e.g. large cards in production on a slow connection that take more than 10 seconds to load). We added retry logic to keep trying to fetch permissions, but that was finicky and also sometimes didn't fire.

The better appraoch is to use onSnapshot, which will fire when the connection is live automatically.

Part of #690.

Coded by Claude.